refactoring_pipelines_finished
